Code cleanup

+++ hardware/ETH8020.cpp
@@ -212,7 +212,7 @@ void CETH8020::GetMeterDetails()
 	}
 	size_t ii;
 	std::string tmpstr;
-	int pos1;
+	size_t pos1;
 	uint8_t Idx = 0;
 	for (ii = 1; ii < results.size(); ii++)
 	{
